requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
docopt==0.6.2
et-xmlfile==1.0.1
fonttools==3.30.0
gevent==1.3.6
greenlet==0.4.15
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jdcal==1.4
lxml==4.2.4
numpy==1.15.1
openpyxl==2.5.6
pandas==0.23.4
parsel==1.5.0
pika==0.12.0
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
python-dateutil==2.7.3
pytz==2018.5
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
Scrapy==1.5.1
scrapy-redis==0.6.8
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.11
stomp.py==4.1.21
threadpool==1.3.2
Twisted==18.7.0
w3lib==1.19.0
zope.interface==4.5.0
